Added Steam and EG default folders for quicker search
Removed the nuget package list as redundant. A checkbox per package seems to be a better idea
Converted the final csproj to new SDK format
Added _Module folder which represents the root folder of the final module output folder
Added version replacing. The version in the SubModule.xml will be taken from csproj
The targets are a bit big. Maybe do something about it
Added better tags for the vstemplate
Added readable template ID
The project will be net472 by default, with enabled nullable and latest language features
Refactored the logic a bit